feat(Issue #10): Implement Proposal and Approval models for Executive and Legislative
- Added Proposal and Approval models to handle EXEC_ACTION and LAW_CHANGE proposals, respectively.
- Implemented validation logic for proposals, including required approvals and payload hash computation.
- Created migrations for the new models and updated serializers and views to integrate with the new structure.
- Removed legacy shared proposals models and adjusted imports across services to reflect the new architecture.
- Enhanced tests to cover the new proposal and approval functionalities, ensuring compliance with the defined business rules.
