+++ services/executive/exec/views.py
@@ -10,14 +10,8 @@ from drf_spectacular.utils import extend_schema, OpenApiParameter
 
 from shared.auth.permissions import RequireScope
 from shared.auth.scopes import PROPOSAL_WRITE, PROPOSAL_FINALIZE, APPROVAL_WRITE
-from shared.proposals.models import (
-    Proposal,
-    Approval,
-    ProposalKind,
-    ProposalOrigin,
-    ProposalStatus,
-    FinalizeConflictError,
-)
+from shared.proposals.common import ProposalKind, ProposalOrigin, ProposalStatus, FinalizeConflictError
+from exec.models import Proposal, Approval
 
 from .models import Evaluation
 from .serializers import EvaluationCreateSerializer, ExecProposalCreateSerializer, ExecApprovalCreateSerializer
