+++ services/executive/executive/settings/base.py
@@ -18,7 +18,6 @@ INSTALLED_APPS = [
     "rest_framework",
     "drf_spectacular",
     "shared.auth",
-    "shared.proposals",
     "exec",
 ]
