--- services/root/approvals/__init__.py
@@ -1,1 +1,0 @@
-# 法則審査系 Approval API（Issue #7）
