--- services/root/approvals/urls.py
@@ -1,6 +1,0 @@
-from django.urls import path
-from . import views
-
-urlpatterns = [
-    path("", views.ApprovalCreateView.as_view(), name="approval-create"),
-]
