--- shared/proposals/apps.py
@@ -1,8 +1,0 @@
-from django.apps import AppConfig
-
-
-class ProposalsAppConfig(AppConfig):
-    default_auto_field = "django.db.models.BigAutoField"
-    name = "shared.proposals"
-    label = "shared_proposals"
-    verbose_name = "Proposal（意思決定最小単位）"
